requirements.txt
streamlit
pandas
matplotlib
supabase
python-dotenv